Stuck in while loop
At some point in the program the device table has all devices free,
therefore completionTimes are all 0, and all IOCompletionTimes are also
all 0. That makes all *LOW* values equal. The way the program is
implemented causes an error when this happens. If all *LOW* times are
equal, no values will change and it will get stuck in the loop.

Currently trying to debug. Going to have to trace the program by hand.